refactor(settings): split ManageAppSettingsContent monolith into section components
Extract each of the 7 settings sections from the 442-line index.tsx into
self-contained components that own their own query/mutation wiring:

- BannerSettingsSection
- TimeZoneSection
- FaqSettingsSection
- DefaultCertificateTemplatesSection
- OnboardingTextsSection

CourseGroupOptionsManager and ProjectDocumentationInstructionsSection were
already extracted; they are unchanged.

Also fixes broken i18n key references in the former monolith: banner, time
zone and FAQ sections were referencing non-existent nested paths
(e.g. `time_zone.label`) instead of the correct flat camelCase keys
(`timeZone`, `timeZoneHelpText`, `timeZones.*`, etc.) that exist in
locales/en.json and locales/de.json.

No UX or behaviour changes.
