+++ frontend-nx/apps/edu-hub/components/pages/ManageAppSettingsContent/index.tsx
@@ -1,442 +1,23 @@
-import { FC, useMemo, useState } from 'react';
-import { FormProvider, SubmitHandler, useForm } from 'react-hook-form';
-import { useSession } from 'next-auth/react';
-import { useTranslations } from 'next-intl';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
+import { FC } from 'react';
 
-import { Button } from '../../common/Button';
-import FormFieldRow from '../../inputs/FormFieldRow';
-import DropDownSelector from '../../inputs/DropDownSelector';
-import CheckboxSelector from '../../inputs/CheckboxSelector';
-import InputField from '../../inputs/InputField';
-
-import { useAdminQuery } from '../../../hooks/authedQuery';
-
-import { UpdateBannerVariables, UpdateBanner } from '../../../queries/__generated__/UpdateBanner';
-import {
-  APP_SETTINGS,
-  UPDATE_APP_SETTINGS_BANNER,
-  UPDATE_APP_SETTINGS_TIME_ZONE,
-  UPDATE_APP_SETTINGS_FAQ_VISIBILITY,
-  UPDATE_APP_SETTINGS_FAQ_COLLECTION,
-  FAQ_COLLECTIONS,
-} from '../../../queries/appSettings';
-import { AppSettings } from '../../../queries/__generated__/AppSettings';
-import { CERTIFICATE_TEMPLATES } from '../../../queries/certificateTemplates';
-import { CertificateTemplates } from '../../../queries/__generated__/CertificateTemplates';
-import {
-  PROGRAM_TYPE_DEFAULTS,
-  UPDATE_PROGRAM_TYPE_DEFAULT_ATTENDANCE_CERTIFICATE_TEMPLATE,
-} from '../../../queries/programTypeDefaults';
-import { ProgramTypeDefaults } from '../../../queries/__generated__/ProgramTypeDefaults';
-import {
-  UpdateProgramTypeDefaultAttendanceCertificateTemplate,
-  UpdateProgramTypeDefaultAttendanceCertificateTemplateVariables,
-} from '../../../queries/__generated__/UpdateProgramTypeDefaultAttendanceCertificateTemplate';
-import { FaqCollections } from '../../../queries/__generated__/FaqCollections';
-import { useAdminMutation } from '../../../hooks/authedMutation';
-import { ONBOARDING_TEXTS, UPDATE_ONBOARDING_TEXT } from '../../../queries/onboardingText';
-import ProjectDocumentationInstructionsSection from './ProjectDocumentationInstructionsSection';
+import BannerSettingsSection from './BannerSettingsSection';
 import CourseGroupOptionsManager from './CourseGroupOptionsManager';
+import DefaultCertificateTemplatesSection from './DefaultCertificateTemplatesSection';
+import FaqSettingsSection from './FaqSettingsSection';
+import OnboardingTextsSection from './OnboardingTextsSection';
+import ProjectDocumentationInstructionsSection from './ProjectDocumentationInstructionsSection';
+import TimeZoneSection from './TimeZoneSection';
 
-type OnboardingTextRow = {
-  id: number;
-  programType: string;
-  lang: string;
-  text: string;
-};
-
-type OnboardingTexts = {
-  OnboardingText: OnboardingTextRow[];
-};
-
-type Inputs = {
-  bannerBackgroundColor: string;
-  bannerFontColor: string;
-  bannerTextDe: string;
-  bannerTextEn: string;
-};
-
-const ManageAppSettingsContent: FC = () => {
-  const { data: sessionData } = useSession();
-  const t = useTranslations('manageAppSettings');
-  const tCourse = useTranslations('course');
-  const tProfile = useTranslations('profile');
-  const [previewProgramType, setPreviewProgramType] = useState('COURSES');
-  const [previewLanguage, setPreviewLanguage] = useState('DE');
-
-  const timeZoneOptions = [
-    { value: 'Europe/Berlin', label: t('time_zone.values.Europe/Berlin') },
-    { value: 'Europe/London', label: t('time_zone.values.Europe/London') },
-    { value: 'Europe/Paris', label: t('time_zone.values.Europe/Paris') },
-    { value: 'UTC', label: t('time_zone.values.UTC') },
-    { value: 'America/New_York', label: t('time_zone.values.America/New_York') },
-    { value: 'America/Los_Angeles', label: t('time_zone.values.America/Los_Angeles') },
-    { value: 'Asia/Tokyo', label: t('time_zone.values.Asia/Tokyo') },
-    // Add more time zones as needed
-  ];
-
-  const methods = useForm<Inputs>({
-    defaultValues: {
-      bannerBackgroundColor: '',
-      bannerFontColor: '',
-      bannerTextDe: '',
-      bannerTextEn: '',
-    },
-  });
-
-  const {
-    handleSubmit,
-    formState: { isSubmitting },
-    reset,
-  } = methods;
-
-  const {
-    loading: appSettingsLoading,
-    error: appSettingsError,
-    data: appSettingsData,
-    refetch: refetchAppSettings,
-  } = useAdminQuery<AppSettings>(APP_SETTINGS, {
-    variables: { appName: 'edu' },
-    onCompleted: (data) => {
-      const [appSettings] = data.AppSettings;
-
-      if (appSettings) {
-        reset({
-          bannerBackgroundColor: appSettings.bannerBackgroundColor ?? undefined,
-          bannerFontColor: appSettings.bannerFontColor ?? undefined,
-          bannerTextDe: appSettings.bannerTextDe ?? undefined,
-          bannerTextEn: appSettings.bannerTextEn ?? undefined,
-        });
-      }
-    },
-    skip: !sessionData,
-  });
-
-  const { data: faqCollectionsData } = useAdminQuery<FaqCollections>(FAQ_COLLECTIONS);
-  const { data: certificateTemplatesData } = useAdminQuery<CertificateTemplates>(CERTIFICATE_TEMPLATES);
-  const { data: programTypeDefaultsData } = useAdminQuery<ProgramTypeDefaults>(PROGRAM_TYPE_DEFAULTS);
-  const { data: onboardingTextsData } = useAdminQuery<OnboardingTexts>(ONBOARDING_TEXTS);
-  const [updateProgramTypeDefaultAttendanceCertificateTemplate] = useAdminMutation<
-    UpdateProgramTypeDefaultAttendanceCertificateTemplate,
-    UpdateProgramTypeDefaultAttendanceCertificateTemplateVariables
-  >(UPDATE_PROGRAM_TYPE_DEFAULT_ATTENDANCE_CERTIFICATE_TEMPLATE, {
-    refetchQueries: ['ProgramTypeDefaults'],
-  });
-
-  const faqCollectionOptions = useMemo(() => {
-    return (
-      faqCollectionsData?.FaqCollection.map((collection) => ({
-        value: collection.name,
-        label: collection.name,
-      })) ?? []
-    );
-  }, [faqCollectionsData]);
-
-  const [updateBanner] = useAdminMutation<UpdateBanner, UpdateBannerVariables>(UPDATE_APP_SETTINGS_BANNER);
-
-  const getOnboardingTextRow = (programType: string, lang: string) =>
-    onboardingTextsData?.OnboardingText.find((row) => row.programType === programType && row.lang === lang);
-
-  const previewText = getOnboardingTextRow(previewProgramType, previewLanguage)?.text ?? '';
-
-  const onSubmit: SubmitHandler<Inputs> = async (data) => {
-    try {
-      await updateBanner({
-        variables: {
-          appName: 'edu',
-          bannerBackgroundColor: data.bannerBackgroundColor,
-          bannerFontColor: data.bannerFontColor,
-          bannerTextDe: data.bannerTextDe,
-          bannerTextEn: data.bannerTextEn,
-        },
-      });
-
-      refetchAppSettings();
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-    } catch (error) {
-      console.error('Failed to update app settings:', error);
-      alert('An error occurred while saving the settings. Please try again.');
-    }
-  };
-
-  return (
-    <div className="px-3 mt-32 max-w-screen-xl mx-auto">
-      {!appSettingsLoading && !appSettingsError ? (
-        <>
-          <div className="mt-16 border border-gray-300 rounded p-6">
-            <label className="text-xs uppercase tracking-widest font-medium text-gray-400 mb-4 block">
-              {t('banner.label')}
-            </label>
-            <FormProvider {...methods}>
-              <form onSubmit={handleSubmit(onSubmit)}>
-                <div className="flex flex-wrap">
-                  <div className="w-1/2 pr-3">
-                    <FormFieldRow<Inputs>
-                      label={t('banner.background_color')}
-                      name="bannerBackgroundColor"
-                      placeholder="#ffffff"
-                      required
-                    />
-                  </div>
-                  <div className="w-1/2 pl-3">
-                    <FormFieldRow<Inputs>
-                      label={t('banner.fontColor')}
-                      name="bannerFontColor"
-                      placeholder="#ffffff"
-                      required
-                    />
-                  </div>
-                </div>
-                <div className="flex flex-wrap">
-                  <div className="w-1/2 pr-3">
-                    <FormFieldRow<Inputs> label={t('banner.text_de')} name="bannerTextDe" />
-                  </div>
-                  <div className="w-1/2 pl-3">
-                    <FormFieldRow<Inputs> label={t('banner.text_en')} name="bannerTextEn" />
-                  </div>
-                </div>
-                <Button
-                  as="button"
-                  type="submit"
-                  disabled={isSubmitting}
-                  filled
-                  inverted
-                  className="mt-8 block mx-auto mb-5 disabled:bg-slate-500"
-                >
-                  {isSubmitting ? t('saving') : t('save')}
-                </Button>
-              </form>
-            </FormProvider>
-          </div>
-
-          <CourseGroupOptionsManager />
-
-          <div className="mt-16">
-            <label className="text-xs uppercase tracking-widest font-medium text-gray-400 mb-2 block">
-              {t('time_zone.label')}
-            </label>
-            <DropDownSelector
-              variant="material"
-              options={timeZoneOptions}
-              value={appSettingsData?.AppSettings[0]?.timeZone ?? ''}
-              helpText={t('time_zone.help_text')}
-              updateValueMutation={UPDATE_APP_SETTINGS_TIME_ZONE}
-              identifierVariables={{ appName: 'edu' }}
-              refetchQueries={['AppSettings']}
-            />
-          </div>
-
-          <div className="mt-16 border border-gray-300 rounded p-6">
-            <label className="text-xs uppercase tracking-widest font-medium text-gray-400 mb-4 block">
-              {t('faq.label')}
-            </label>
-
-            {/* Show FAQ Section Checkbox */}
-            <div className="mb-6">
-              <CheckboxSelector
-                variant="material"
-                label={t('faq.show_section')}
-                checked={appSettingsData?.AppSettings[0]?.showFaqSection ?? true}
-                updateValueMutation={UPDATE_APP_SETTINGS_FAQ_VISIBILITY}
-                identifierVariables={{ appName: 'edu' }}
-                refetchQueries={['AppSettings']}
-                className="text-gray-300"
-              />
-            </div>
-
-            {/* FAQ Collection Dropdown */}
-            <div>
-              <label className="block text-base font-medium text-gray-300 mb-3">{t('faq.collection_name')}</label>
-              <DropDownSelector
-                variant="material"
-                options={faqCollectionOptions}
-                value={appSettingsData?.AppSettings[0]?.faqCollectionName ?? 'default'}
-                placeholder={t('faq.collection_placeholder')}
-                updateValueMutation={UPDATE_APP_SETTINGS_FAQ_COLLECTION}
-                identifierVariables={{ appName: 'edu' }}
-                refetchQueries={['AppSettings']}
-              />
-            </div>
-          </div>
-
-          <div className="mt-16 border border-gray-300 rounded p-6">
-            <label className="text-xs uppercase tracking-widest font-medium text-gray-400 mb-4 block">
-              {t('default_attendance_certificate_template.label')}
-            </label>
-            <p className="text-xs text-gray-400 mb-4">
-              {t('default_attendance_certificate_template.help_text')}
-            </p>
-            <div className="space-y-4">
-              {(programTypeDefaultsData?.ProgramType ?? []).map((row) => (
-                <div key={row.value}>
-                  <label className="block text-base font-medium text-gray-300 mb-2">
-                    {t(`programTypes.${row.value}`)}
-                  </label>
-                  <select
-                    className="block w-full text-sm rounded border border-border-primary bg-fill-primary p-2"
-                    value={row.defaultAttendanceCertificateTemplateId ?? ''}
-                    onChange={(e) =>
-                      updateProgramTypeDefaultAttendanceCertificateTemplate({
-                        variables: {
-                          value: row.value,
-                          templateId: e.target.value === '' ? null : parseInt(e.target.value, 10),
-                        },
-                      })
-                    }
-                  >
-                    <option value="">{t('default_attendance_certificate_template.none_option')}</option>
-                    {(certificateTemplatesData?.CertificateTemplate ?? []).map((tpl) => (
-                      <option key={tpl.id} value={tpl.id}>
-                        {tpl.name}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          <ProjectDocumentationInstructionsSection />
-
-          <div className="mt-16 border border-gray-300 rounded p-6">
-            <label className="text-xs uppercase tracking-widest font-medium text-gray-400 mb-4 block">
-              {t('onboarding.label')}
-            </label>
-
-            {(['COURSES', 'EVENTS', 'DEGREES'] as const).map((programType) => {
-              let titleKey = 'onboarding.degrees_label';
-              if (programType === 'COURSES') {
-                titleKey = 'onboarding.courses_label';
-              } else if (programType === 'EVENTS') {
-                titleKey = 'onboarding.events_label';
-              }
-              const deText = getOnboardingTextRow(programType, 'DE');
-              const enText = getOnboardingTextRow(programType, 'EN');
-
-              return (
-                <div key={programType} className="mb-10">
-                  {deText && enText ? (
-                    <div className="max-w-4xl">
-                      <InputField
-                        variant="eduhub"
-                        type="markdown"
-                        label={t(titleKey)}
-                        helpText={t('onboarding.generic_help')}
-                        translationTabs={[
-                          { lang: 'DE', itemId: deText.id, value: deText.text },
-                          { lang: 'EN', itemId: enText.id, value: enText.text },
-                        ]}
-                        itemId={deText.id}
-                        value={deText.text}
-                        updateValueMutation={UPDATE_ONBOARDING_TEXT}
-                        refetchQueries={['OnboardingTexts']}
-                        maxLength={10000}
-                        className="h-64"
-                      />
-                    </div>
-                  ) : (
-                    <div className="text-sm text-label-secondary">Loading...</div>
-                  )}
-                </div>
-              );
-            })}
-
-            <div>
-              <label className="text-xs uppercase tracking-widest font-medium text-gray-400 mb-3 block">
-                {t('onboarding.preview')}
-              </label>
-              <div className="flex flex-col lg:flex-row gap-3 mb-4">
-                <select
-                  className="rounded border border-border-primary px-3 py-2 bg-fill-primary text-label-primary"
-                  value={previewProgramType}
-                  onChange={(event) => setPreviewProgramType(event.target.value)}
-                >
-                  <option value="COURSES">{t('onboarding.courses_label')}</option>
-                  <option value="EVENTS">{t('onboarding.events_label')}</option>
-                  <option value="DEGREES">{t('onboarding.degrees_label')}</option>
-                </select>
-                <select
-                  className="rounded border border-border-primary px-3 py-2 bg-fill-primary text-label-primary"
-                  value={previewLanguage}
-                  onChange={(event) => setPreviewLanguage(event.target.value)}
-                >
-                  <option value="DE">{t('onboarding.text_de')}</option>
-                  <option value="EN">{t('onboarding.text_en')}</option>
-                </select>
-              </div>
-              <div className="max-w-4xl">
-                <div className="bg-edu-course-invited rounded-2xl p-6 text-label-primary light border border-border-primary/30 shadow-lg">
-                  <div className="mb-6 rounded-xl bg-fill-primary p-4">
-                    <ReactMarkdown
-                      className="prose max-w-none text-label-primary prose-headings:font-bold prose-headings:text-label-primary prose-p:text-label-secondary prose-h1:text-2xl prose-h2:text-xl prose-h3:text-lg"
-                      remarkPlugins={[remarkGfm]}
-                    >
-                      {previewText}
-                    </ReactMarkdown>
-                  </div>
-
-                  <div className="flex flex-wrap gap-y-3 mb-3">
-                    <div className="w-full lg:w-1/2 lg:pr-3">
-                      <label className="block text-sm text-label-primary mb-2">{tProfile('occupation.label')}</label>
-                      <input
-                        type="text"
-                        value={tProfile('occupation.UNIVERSITY_STUDENT')}
-                        readOnly
-                        disabled
-                        className="w-full px-3 py-3 rounded bg-fill-primary text-label-primary border border-border-primary opacity-80"
-                      />
-                    </div>
-                    <div className="w-full lg:w-1/2 lg:pl-3">
-                      <label className="block text-sm text-label-primary mb-2">{tProfile('organization.label_university')}</label>
-                      <input
-                        type="text"
-                        value={tProfile('organization.placeholder')}
-                        readOnly
-                        disabled
-                        className="w-full px-3 py-3 rounded bg-fill-primary text-label-primary border border-border-primary opacity-80"
-                      />
-                    </div>
-                  </div>
-
-                  <div className="w-full lg:w-1/2 lg:pr-3 mb-2">
-                    <label className="block text-sm text-label-primary mb-2">{tProfile('matriculation_number')}</label>
-                    <input
-                      type="text"
-                      value=""
-                      readOnly
-                      disabled
-                      className="w-full px-3 py-3 rounded bg-fill-primary text-label-primary border border-border-primary opacity-80"
-                    />
-                  </div>
-
-                  <div className="flex flex-col lg:flex-row lg:gap-5">
-                    <button
-                      type="button"
-                      disabled
-                      className="mt-8 block mx-auto lg:mb-5 px-8 py-3 rounded-full bg-error text-label-primary border-2 border-error opacity-100"
-                    >
-                      {tCourse('general.reject')}
-                    </button>
-                    <button
-                      type="button"
-                      disabled
-                      className="mt-4 lg:mt-8 block mx-auto lg:mb-5 px-8 py-3 rounded-full bg-success text-label-primary border-2 border-success opacity-100"
-                    >
-                      {tCourse('general.confirm')}
-                    </button>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </>
-      ) : (
-        <div>Loading</div>
-      )}
-    </div>
-  );
-};
+const ManageAppSettingsContent: FC = () => (
+  <div className="px-3 mt-32 max-w-screen-xl mx-auto">
+    <BannerSettingsSection />
+    <CourseGroupOptionsManager />
+    <TimeZoneSection />
+    <FaqSettingsSection />
+    <DefaultCertificateTemplatesSection />
+    <ProjectDocumentationInstructionsSection />
+    <OnboardingTextsSection />
+  </div>
+);
 
 export default ManageAppSettingsContent;
